requirements.txt
requests
pyyaml
python-bitcoinlib
tenacity
